Made server and client application standalone
