+++ specs/Naludrag/common/Protocol.java
@@ -1,5 +1,4 @@
 package common;
-
 /**
  *
  * @author Olivier Liechti
