+++ specs/Naludrag/server/ServerApplication.java
@@ -1,0 +1,9 @@
+package server;
+
+public class ServerApplication {
+    public static void main(String[] args) {
+        Server srv = new Server();
+        srv.run();
+        //srv.shutdown();
+    }
+}
